Use internal subclass to hide Decoder constructor
Mark the constructor private in the exported Decoder class and create an
InternalDecoder subclass that publicizes the constructor for internal
use.

We now know that a consumers of this library can only create a Decoder
by using our exported functions, or by subclassing Decoder and hopefully
assuming the risk of depending on undocumented functionality.